Add player heads and mob skulls
So this makes a 1x8x8 cube for all mobs, and then maps the mob skin onto
the head cube, so that skulls better fit into our flat textures.

With exceptions for Piglins (that have a 10x8 skull) and Players. We’re
unable to make a custom model for players as their texture isn’t passed
to models, so instead we use scaling to make the heads kinda 1 cube
deep. This makes the sides of heads distorted, but it’s the best of what
we can do right now. player_head_pixel allows non-distorted heads, but
no custom skins so just steve.

Yet, I seem unable to refer to vanilla mob entity textures inside
models, so had to include them in this pack for now. Also since I am
still new to modelling for Minecraft, mob skins without the overlay
section get an incorrect mapping (UVs scale to texture size), so until I
know how to do absolute texture references, adjusted those textures to
all be 64×64.

Basically that is all to say, this is a WIP and if I discover better
ways to do any of that, it’ll be done in the future.