Fix shield model
Another oversight that I didn’t notice the original project had, and
removed the blocking shield override as we’re just focused on shulker
labelling, and adjusted the scaling and translation of the fixed display
of shields - which roughly makes the shield fit our aesthetic without
model changes.

Though a bug in Peek makes it sub par. (everything is 180° rotated)