Add respackopts as an optional dependency
A lot of new files, that are completely ignorable to users without the
respackopt mod. It allows users that wish to have no animations, or
emissive textures, to disable them.

It’s a great mod for resourcepacks to have minor configurable
behaviours, and I hope that this wont expand much further than this.

I may have missed the odd animation, and I need to make a quick terminal
lookup for any non-square textures, but otherwise this should be it.

Had written out a fairly lengthy description of why bits and bobs were
done (_static.png, _e_static.mcmeta, and all the rpo files) but my GPG
key had expired so git decided to eat that commit msg. Sigh.

Updated and added screenshots for the more recent updates, and once I
make something to generate shots of supported blocks, long.png will be
updated finally