Remove unused textures
This repo included the hilarious obamite texture that was part of the
release files from https://github.com/ThePotatoKing55/2D-block-texture-pack/releases/tag/v0.8-alpha.1
which was the basis of a test that developed into this pack. It’s a
cute easter egg, but unlinked in any of the models, so less just retire
it. grass_block was converted to use overlays, as were suspicious_*
textures. target_variant was held in case the block appearance wasn’t
good enough, and I think it’s time that goes now too. sculk_sensor got
replaced with a newer texture, so it goes aswell.

There are still some remaining, and I am debating if they need to stay.
* comparator and repeater seem the most desirable.
* sea_pickle, chain, glistering_melon_slice should get the chop
* end_crystal may need to be kept, for emissive purposes to match others