Added Polished Tuff (& variants), Tuff Bricks (& variants), Chiseled too
With this, update the regular Tuff appearance to have a more muted edge,
closer to the average colour of the block.

Very much something made by a non-artist, so this may have tweaks in
future. Chiseled tuff and chiseled tuff bricks require a bit of
squinting as they are quite complex patterns that have to fit into ~12
pixels or so.