Added all variants of Copper trapdoors
Like the previous commit, very much something made by a non-artist, so
this may have tweaks in future.

Made the waxed versions appear the same as the unwaxed, for now. I like
the infested blocks having a whitish border, and the vanillatweaks
resourcepack I rock normally has waxed blocks with a yellow border, so
am thinking that is the likely thing we do in future.

Got half-way into cut_copper_trapdoor model files and then checked the
wiki. That’s an oddity to not have a completely different looking copper
trapdoor, would be neat.