Add Copper grates
Using generated models for this, but we could easily have used a custom
model and then applied the base copper textures (like we do for slabs,
stairs, walls, etc) but it’s a hole-y texture, which doesn’t make for a
very compact model. Can change in future if more grate materials appear.

Made the waxed versions appear the same as the unwaxed, for now.