Add all variants of Chiseled Copper
Very much something made by a non-artist, so this may have tweaks in
future. Or be completely redone.

Thinking that in future we bake overlay stuff down. It makes for a less
flexible resourcepack, but removes half the runtime cruft and should
make atlases smaller - which I will be looking at before we get to v1.0

It also makes for easier texture creation, as these overlays are just
muddier layers on the base textures. Since we cannot use alpha blending
(the two layers are not combined by the game), it makes us just bake the
lower texture into the overlay.

This puts us at needing Copper Bulbs, Crafter, and Trial Spawners before
we’re good for 1.21.