Add Copper Bulb, Crafter, Trial Spawner & Vault blocks
Unless there are changes with new blocks arriving in 1.21, or major
texture changes, we will be up to date with the new version when it
launches. That’s not to say we are done, I still have further tweaks to
make this, imo, a good texture pack for vanilla play while accepting a
few quality-of-life mods. Of course, there are wax variants of the bulb,
which I imagine are going to be 99% of the use of the thing, so maybe we
do want to work on that wax overlay before we say we’re done.

Further digging into the emissive issue points to Continuity not seeing
the emissive textures, due to the reasons mentioned in previous commit.
I will look into it more and see if, we can do something here, or file
an issue report with that Sodium additive.

Trial spawner and vault textures have change a few times between recent
snapshots, so hoping they don’t change again.

Every models/item is from 1.20.4, with the exception of vault.json,
which apparently is still only in snapshot jars.