Add waxed trapdoors
Following previous changes, we’re adding honey rims to all the waxed
blocks, so this just leaves doors as unmarked.

This new overlay_template uses the trapdoor texture I made, and the
waxed block border that has been used elsewhere. We could remap the UVs
while reusing the regular copper block variants to further reduce the
amount of various copper related textures that are made. Maybe in a
future update.