Add waxed doors
Not overly happy with the appearance of this right now, mostly ’cause
the shading needs to be better, but I don’t want another overlay texture
just for doors.

All of this is mostly moot, ’cause who is going to be putting stacks of
waxed trapdoors or doors into their shulkers? I suppose to uniquely
label a storage container, ‘{i=waxed_copper_door}’ might be your choice.
But it would be a poor one :D

With this commit, every new block _and variant_ for 1.21 is included in
the pack. Since the last release I went through and checked rendering
for a lot of items since that Peek rotation bug was fixed, improved the
appearance of many items, introduced the waxed overlay, made new blocks,
removed redundant files, and even resolved a problem I introduced..