Release v0.8
Nearly ready for 1.0, this is noew a functional resource pack for Peek,
making shulkers easily identifiable in your inventory or placed in the
world.

It might be odd to list the TODOs in the release, but I think I want to
commit to future next changes (in case there is feedback) here.

* Buttons don’t fit, they need to be smaller and centrally aligned.
* Stained glass and panes are hard to see on same colour shulkers
* Bed colours need slight adjustments
* Compile a replacement long.png of supported blocks